Enable 32 column tables by default, move features around
This deprecates the names `large-tables` and `huge-tables` in favor of
`32-column-tables` and `64-column-tables`. `32-column-tables` is enabled
by default. On my machine this changes the compile time of Diesel from 8
seconds to 18 seconds, which I think is an acceptable loss. The feature
can still be turned off for cases where compile times are important.