Rewrite `#[derive(AsExpression)]` in derives2
This rewrite was pretty straightforward. Getting the span correct on the
type required a bit of work. I'm not sure if there's a better way to do
this or not (perhaps syn just needs a `parse_str_span` function).

This is also the first derive that's used by Diesel itself to get moved.
This required some funkyness in the module wrapping. If
`__diesel_use_everything!()` has `Span::def_site()`, for some reason
that macro doesn't resolve when used inside of Diesel (but it works fine
when used from a third crate). However, if the `;` after it *doesn't*
have `Span::def_site()`, none of the imports are visible to the macro...
This was fun to figure out.

This derive is pretty lightly tested, since we use it for *everything*
in Diesel itself, and that leads to a pretty exhaustive test that way.
This macro also has far fewer error cases than "model" derives.