Rewrite `#[derive(Insertable)]` in derives2
I didn't realize just how much we'd simplified this derive at this
point. This used to be by far our most involved derive.

I've explicitly opted not to include a UI test for when there's a type
mismatch, as I cannot get errors that I'm satisfied with yet, and I
suspect that it will have to wait for some changes in the language.