Rewrite `#[derive(QueryId)]` in derives2
This was a really weird one to get working with hygiene turned on. I'm
not sure if this is a bug in rustc, or something we need to explicitly
deal with. Until we know for sure, we need to do this hack to make this
work. At the moment proc_macro2 doesn't expose `Span::source` even with
`procmacro2_semver_exempt`.